+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Features
 
 - usbip: Add `--efs` option to store the external filesystem in a file.
+- Add variant to the status reported by admin-app ([#206][])
+
+[#206]: https://github.com/Nitrokey/nitrokey-3-firmware/issues/206
 
 # v1.3.1 (2023-04-05)